Have to pause on getting windshaft testable
I've got the environment for test buildable in terms of the test
databases, but the tests themselves are difficult to run successfully
because the devs hardcoded a number of things into the tests and/or
test helpers. Going to pause for the time being, return to this after
getting the Carto container working.