Getting close to having cartodb container working
Frustratingly, the `bundle exec rake db:create` step is failing, likely
due to a bug in the `sequel-rails` package:

    https://github.com/TalentBox/sequel-rails/issues/158

The workaround seems to be manually creating the database, because
sequel-rails screws something up such that db:create fails due to the
database not existing. Something in the initializers about attempting to
connect to the db prior to actually creating it, but I wasn't able to
ferret it out or successfully implement their supposed fixes.