Egregious hackery to allow HTTPS in dev
Their check for whether to use https in URL construction is based on the
Rails env being either production or staging. I added a bit to the
docker-entrypoint.sh script that will edit in place the use_https?
method in the app initializer, so it just returns true. It only does the
edit if there's a corresponding 'use_https: true' entry in
app_config.yml, though that setting is one I made up, not an official
Carto one.