Moved to http based varnish purge
Commenting out the `invalidation_service` section of the app_config.yml
file causes the app to set a postgres trigger for varnish purges based
not on the resque task queue, but a direct call to Varnish over http to
purge a specific cache item. This is a stop gap until I can determine
whether it is possible to run the invalidation service without the
custom Redis commands that queue related resque jobs.